Extract parse-title and youtube-search business cores
