+++ _api/_runtime/core-types.ts
@@ -1,0 +1,5 @@
+export interface CoreResponse {
+  status: number;
+  body: unknown;
+  headers?: Record<string, string>;
+}
